+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   * Activity class - versions now returns an array of Osm::Activity::Version objects
   * Section class - rename extra\_records to flexi\_records
   * Section class - flexi_records now returns an array of Osm::Section::FlexiRecord objects
+  * get\_register becomes get\_register\_data and now returns an array of RegisterData objects
+  * get\_register\_structure now returns an array of RegisterField objects
 
 ## Version 0.0.11